Add population endpoints and update City model
Updated CountriesController to include endpoints for country population data. Introduced CountryPopulation DTO for structured data. Changed City model's Population property from int to decimal for precision. Created migrations to update the database schema accordingly. Added CitiesController for CRUD operations on cities.

+++ WorldModel/City.cs
@@ -28,7 +28,7 @@ public partial class City
     public decimal Longitude { get; set; }
 
     [Column("population")]
-    public int Population { get; set; }
+    public decimal Population { get; set; }
 
     [ForeignKey("CountryId")]
     [InverseProperty("Cities")]
